Remove the icon (nullPointerException) and move the Utils folder into an singe file
